change the packet font from Courier to Hack
Courier is a monospace font that has ligatures, i.e., it can convert some
combinations of characters such as "ff", "fi", "fl",... to single characters.
While this is visualy nice most of the time, this misaligns packet format
pictures.

Change it to the Hack font, which has no ligature and is also more readable to
me.

Fixes #32.